requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
cffi==1.16.0
cryptography==41.0.7
decorator==5.1.1
Deprecated==1.2.14
Django==4.2.7
djangorestframework==3.12.4
fabric==3.2.2
fake==0.8
invoke==2.2.0
paramiko==3.4.0
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyNaCl==1.5.0
sqlparse==0.4.4
tzdata==2023.3
wrapt==1.16.0
faker-22.2.0